% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsmm2hmm2.R
\name{hsmm2hmm2}
\alias{hsmm2hmm2}
\title{Converting hsmm to hmm (sparse)}
\usage{
hsmm2hmm2(omega, dm, eps = 1e-10)
}
\arguments{
\item{omega}{Matrix of conditional transition probabilities}

\item{dm}{list (of states) of pmf vectors}

\item{eps}{accuracy parameter}
}
\value{
Returns sparse Gamma matrix
}
\description{
Converting hsmm to hmm (sparse)
}
