% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phsmm2phmm.R
\name{phsmm2phmm}
\alias{phsmm2phmm}
\title{Converting periodic hsmm to periodic hmm}
\usage{
phsmm2phmm(omega, dm, eps = 1e-25)
}
\arguments{
\item{eps}{}
}
\description{
Converting periodic hsmm to periodic hmm
}
