% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phsmm2phmm2.R
\name{phsmm2phmm2}
\alias{phsmm2phmm2}
\title{Converting periodic hsmm to periodic hmm}
\usage{
phsmm2phmm2(omega, dm, eps = 1e-25)
}
\arguments{
\item{omega}{Matrix of conditional transition probabilities}

\item{dm}{a list (in time) of lists (of states) of pmf vectors}

\item{eps}{accuracy parameter}
}
\value{
List of Sparse Gamma matrices
}
\description{
Converting periodic hsmm to periodic hmm
}
