% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pv.R
\name{pv}
\alias{pv}
\title{Convenient computation of trigonometric predictor}
\usage{
pv(coef, time, degree, L = 24)
}
\arguments{
\item{coef}{Matrix of beta coefficients}

\item{time}{timepoint to evaluate}

\item{degree}{number of trigonometric functions to use}

\item{L}{length of one period}
}
\description{
Convenient computation of trigonometric predictor
}
