% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsmm2hmm.R
\name{hsmm2hmm}
\alias{hsmm2hmm}
\title{Converting hsmm to hmm}
\usage{
hsmm2hmm(omega, dm, eps = 1e-10)
}
\arguments{
\item{omega}{Matrix of conditional transition probabilities}

\item{dm}{list (of states) of pmf vectors}

\item{eps}{accuracy parameter}
}
\description{
Converting hsmm to hmm
}
